chore: disable file upload
